+++ packages/create-wave-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-wave-app
 
+## 0.15.0
+
+### Minor Changes
+
+- Creates deploy service, refactors api gateway authorizer, writes sst deployment command
+
 ## 0.14.1
 
 ### Patch Changes