+++ services/api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # api-service
 
+## 0.1.2
+
+### Patch Changes
+
+- @vastly/utils@0.3.2
+
 ## 0.1.1
 
 ### Patch Changes